atc-brain-python/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
asyncpg==0.30.0
redis==6.4.0
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0
httpx==0.28.1
pyyaml==6.0.3

